Change userIdx to applicantIdx in all applicant action. Change url param (:applicantId -> applicantIdx) to clarify meaning.

+++ controller/applications.js
@@ -14,15 +14,15 @@ const updateApplication = async (req) => {
   const t = await models.sequelize.transaction();
 
   try {
-    const userIdx = req.params.applicantId;
-
+    const applicantIdx = req.user.applicantIdx;
+    const applicantInfo = await models.applicantInfoTb.findOne(
+      { where: { applicantIdx }, transaction: t });
+    const userIdx = applicantInfo.userIdx;
     const data = req.body;
-    const userName = data.userName;
-    const userPosition = data.userPosition;
+    const { userName, userPosition } = data;
     data.interviewAvailableTime.forEach((elem, i, arr) => { arr[i] = new Date(elem); });
 
     const applicantData = {
-      userIdx,
       applicantGender: data.applicantGender,
       applicantBirthday: new Date(data.applicantBirthday),
       applicantLocation: data.applicantLocation,
@@ -33,7 +33,7 @@ const updateApplication = async (req) => {
     };
 
     const appDocData = {
-      userIdx,
+      applicantIdx,
       entryRoute: data.entryRoute,
       portfolioFileUrl: data.portfolioFileUrl,
       personalUrl: data.personalUrl,
@@ -45,9 +45,9 @@ const updateApplication = async (req) => {
     const userInfoResult = await models.userInfoTb.update({ userName, userPosition },
       { where: { userIdx }, transaction: t });
     const applicantInfoResult = await models.applicantInfoTb.update(applicantData,
-      { where: { userIdx }, transaction: t });
+      { where: { applicantIdx }, transaction: t });
     const appDocResult = await models.applicationDoc.findOneAndUpdate(
-      { userIdx }, appDocData, { runValidators: true });
+      { applicantIdx }, appDocData, { runValidators: true });
     await t.commit();
     // 결과값 응답
     return [userInfoResult, applicantInfoResult, appDocResult];
@@ -88,13 +88,13 @@ module.exports.submitApplication = async (req, res, next) => {
 // Form에 줄 정보만 선별적으로 만들어야 해서 Join 연산보다 쿼리 두번 실행하는 것이 낫다고 판단
 module.exports.getMyApplication = async (req, res, next) => {
   try {
-    const userIdx = Number(req.params.applicantId);
-    const applicationDocRet = await models.applicationDoc.findOne({ userIdx }).exec();
+    const applicantIdx = Number(req.params.applicantIdx);
+    const applicationDocRet = await models.applicationDoc.findOne({ applicantIdx }).exec();
+    const applicantInfo = await models.applicantInfoTb.findOne({ where: { applicantIdx } });
+    const applicantInfoRet = applicantInfo.dataValues;
+    const userIdx = applicantInfoRet.userIdx;
     const userInfo = await models.userInfoTb.findOne({ where: { userIdx } });
     const userInfoRet = userInfo.dataValues;
-    const applicantInfo = await models.applicantInfoTb.findOne({ where: { userIdx } });
-    const applicantInfoRet = applicantInfo.dataValues;
-
     const result = {
       // From UserInfo
       userName: userInfoRet.userName,
@@ -123,14 +123,13 @@ module.exports.getMyApplication = async (req, res, next) => {
 
 module.exports.removeApplication = async (req, res, next) => {
   try {
-    const userIdx = req.params.applicantId;
+    const applicantIdx = req.params.applicantIdx;
     const result = [];
-    const data = await models.applicantInfoTb.findOne({ where: { userIdx } });
-    result.push(await models.applicantInfoTb.destroy({ where: { userIdx } }));
-    result.push(await models.userInfoTb.destroy({ where: { userIdx } }));
-    result.push(await models.applicationDoc.remove({ userIdx }));
-    result.push(await models.applicationTb.destroy(
-      { where: { applicantIdx: data.dataValues.applicantIdx } }));
+    const data = await models.applicantInfoTb.findOne({ where: { applicantIdx } });
+    result.push(await models.applicantInfoTb.destroy({ where: { applicantIdx } }));
+    result.push(await models.userInfoTb.destroy({ where: { userIdx: data.dataValues.userIdx } }));
+    result.push(await models.applicationDoc.remove({ applicantIdx }));
+    result.push(await models.applicationTb.destroy({ where: { applicantIdx } }));
     res.r(result);
   } catch (err) {
     next(err);
